// client/src/model/IMovieList.ts
export default interface IMovieList {
    id: string;
    title: string;
    year: string;
    genres: string[];
    ratings: number[];
    poster: string;
    contentRating: string;
    duration: string;
    releaseDate: string;
    averageRating: number;
    originalTitle: string;
    storyline: string;
    actors: string[];
    imdbRating: string;
    posterurl: string;
}